+++ lib/report_desk_bootstrap.lua
@@ -2,6 +2,12 @@
 require 'lib.moonloader'
 require 'lib.sampfuncs'
 
+pcall(function()
+    if package.loaded['lib.sampfuncs'] and not package.loaded['sampfuncs'] then
+        package.loaded['sampfuncs'] = package.loaded['lib.sampfuncs']
+    end
+end)
+
 local function ensureIconvDll()
     local iconvPath = getWorkingDirectory() .. '\\lib\\iconv.dll'
     if doesFileExist(iconvPath) then